changed int to long int

+++ 0x0C-more_malloc_free/101-mul.c
@@ -8,9 +8,9 @@
  */
 int main(int argc, char *argv[])
 {
-	int mul;
-	int num1 = atoi(argv[1]);
-	int num1 = atoi(argv[2]);
+	long int mul;
+	long int num1 = atoi(argv[1]);
+	long int num1 = atoi(argv[2]);
 
 	if (argc != 3)
 	{
